requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-plugins==1.1.1.2
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
dotenv==0.9.9
fritz==2.3.5
idna==3.10
MarkupSafe==3.0.2
openstep_parser==2.0.1
pbxproj==4.3.0
psycopg2-binary==2.9.10
pybind11==3.0.0
PyJWT==2.10.1
PyMuPDF==1.26.1
PyPDF2==3.0.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
termcolor==3.1.0
urllib3==2.5.0
Werkzeug==3.1.3
